+++ lib/TransactWebpage.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:webview_flutter_plus/webview_flutter_plus.dart';
@@ -22,10 +24,35 @@ class _TransactWebpageState extends State<TransactWebpage> {
   List<RespDataModel> resSplitData = [];
 
   Future<bool> _onBackPressed() async {
+  //   print("In TRANSACT WEB ");
+  //   ResponseConfig.startTrxn = false;
+  //   Navigator.pop(context, null);
+  // //  Navigator.of(context).pop(true);
+  //   return Future.value(false);
 
-    ResponseConfig.startTrxn = false;
-    Navigator.of(context).pop(true);
-    return Future.value(false);
+    bool willLeave = false;
+    // show the confirm dialog
+    await showDialog(
+        context: context,
+        builder: (_) => AlertDialog(
+          title: const Text('Are you sure?'),
+          content: const Text("Do you want to go back"),
+          actions: [
+            TextButton(
+                onPressed: () => Navigator.pop(context, null),
+                child: const Text('No')),
+            TextButton(
+                onPressed: () {
+                  willLeave = true;
+                  Navigator.pop(context, null);
+                  ResponseConfig.startTrxn = false;
+
+                },
+                child: const Text('Yes')),
+
+          ],
+        ));
+    return willLeave;
   }
 
 
@@ -37,8 +64,15 @@ class _TransactWebpageState extends State<TransactWebpage> {
     return new WillPopScope(
       onWillPop: _onBackPressed,
       child:
-      SafeArea(
-          child: Container(
+      Scaffold(
+          appBar: AppBar(
+            backgroundColor: Colors.white,
+            iconTheme: IconThemeData(
+              color: Colors.black, // <-- SEE HERE
+            ),
+            title: const Text('Payment', style: TextStyle(color: Colors.black)),
+          ),
+          body: Container(
             margin: const EdgeInsets.all(1.0),
             child: WebViewPlus(
               initialUrl: widget.inURL,
@@ -55,7 +89,7 @@ class _TransactWebpageState extends State<TransactWebpage> {
               },
 
               onPageFinished: (url) {
-                // print(this.url)
+               // print(this.url);
                 var disurl = url.toString();
                 if (disurl.contains("&Result")) {
                   List<String> arr = url.toString().split('?');
@@ -63,14 +97,15 @@ class _TransactWebpageState extends State<TransactWebpage> {
                   // print('RES DATA $resData');
                   //String lastData=splitResponse(arr[1]);
 
-                  List<RespDataModel> mapData = RespJson(arr[1]);
-                  // print('Transact List $mapData');
-
-
-                  var map1 = Map.fromIterable(
-                      mapData, key: (e) => e.resKey, value: (e) => e.resValue);
+                 String mapData = RespJson(arr[1]);
+                // print(mapData);
+                  // // print('Transact List $mapData');
+                  //
+                  //
+                  // var map1 = Map.fromIterable(
+                  //     mapData, key: (e) => e.resKey, value: (e) => e.resValue);
                   // print(map1);
-                  Navigator.pop(context, '$map1');
+                  Navigator.pop(context, '$mapData');
                 }
 
               },
@@ -79,16 +114,14 @@ class _TransactWebpageState extends State<TransactWebpage> {
                   _progress = (progress / 100) as double;
                 });
               },
-
-
             ),
           )
       ),
     );
   }
-    List<RespDataModel> RespJson(String resultData) {
+   String RespJson(String resultData) {
       List<String> resultParameters = resultData.split("&");
-
+      var responseData = {};
 
       for(final params in resultParameters){
 
@@ -101,16 +134,44 @@ class _TransactWebpageState extends State<TransactWebpage> {
           value= "";
         }
 
+         // metadata["entryone"] = metadata1;
+
+        if("metaData".contains(parts[0]))
+          {
+            String strdecoded=parts[1];
+         //   print(strdecoded);
+            // String decoded = utf8.decode(base64.decode(strdecoded));
+            // print(decoded);
+
+            String strMetadata = strdecoded.split('.')[0];
+            List<int> res = base64.decode(base64.normalize(strMetadata));
+
+           String metadata = utf8.decode(res);
+
+            resSplitData.add(RespDataModel(resKey: "metaData", resValue: metadata));
+            responseData["metaData"] = metadata;
+          }
+        else{
+        //  print(value);
+          resSplitData.add(RespDataModel(resKey: key, resValue: value));
+          responseData[key] = value;
+        }
+
         var productMap = {
 
           key: value,
 
         };
 
-        resSplitData.add(RespDataModel(resKey: key, resValue: value));
+      //  resSplitData.add(RespDataModel(resKey: key, resValue: value));
       }
+
+      String metaJson = json.encode(responseData);
+      // String metaJson = encoder.convert(metadata);
+     // print('Json REsponse Data $metaJson');
       ResponseConfig.startTrxn = false;
-      return resSplitData;
+    //  print('REsponse  $resSplitData ');
+      return metaJson;
 
 
     }